requirements.txt
ansimarkup==1.4.0
beautifulsoup4==4.6.3
better-exceptions-fork==0.2.1.post6
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
fire==0.1.3
idna==2.8
loguru==0.2.3
lxml==4.2.5
numpy==1.11.3
pandas==0.20.1
Pygments==2.3.1
pymc==2.3.6
pymongo==3.7.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==4.2b1
requests==2.21.0
scipy==0.19.0
six==1.10.0
tornado==5.1.1
urllib3==1.24.1
